+++ packages/cli/src/workflows/create/Create.common.handler.ts
@@ -9,7 +9,7 @@
 *
 */
 
-import { IHandlerParameters, ImperativeError } from "@zowe/imperative";
+import { Censor, IHandlerParameters, ImperativeError } from "@zowe/imperative";
 import {
     ListWorkflows,
     DeleteWorkflow,
@@ -185,7 +185,7 @@ export default class CreateCommonHandler extends ZosmfBaseHandler {
                     msg:
                         `Internal create error: Unable to determine the source of the definition file. ` +
                         `Please contact support.`,
-                    additionalDetails: JSON.stringify(params)
+                    additionalDetails: JSON.stringify(Censor.censorObject(params.arguments))
                 });
         }
     }